refactor: Export constants and classes in index.ts
This commit refactors the code in index.ts to export the constants and classes as named exports. Previously, these were defined as constants and classes within the file, but now they are exported for use in other modules.

The changes include:
- Exporting constants SEOUL_LAT_MIN, SEOUL_LAT_MAX, SEOUL_LNG_MIN, SEOUL_LNG_MAX, MOVE_DISTANCE_MIN, MOVE_DISTANCE_MAX, TOPIC, INTERVAL_MAX_MS, INTERVAL_MIN_MS, and COUNT.
- Exporting the interfaces MobilityInfo.
- Exporting the classes Mobility and MobilityManager.

These changes improve the modularity and reusability of the codebase.